+++ tests/FSharp.AWS.DynamoDB.Tests/Program.fs
@@ -201,6 +201,17 @@ let SparseGSITests =
             testCase "GSI Query Operation (missing)" sparseGSITests.``GSI Query Operation (missing)``
         ]
 
+let paginationTestsTableFixture = new TableFixture()
+let paginationTests = ``Pagination Tests``(paginationTestsTableFixture)
+let PaginationTests =
+    testList "PaginationTests"
+        [
+            testCase "Paginated Query on Primary Key" paginationTests.``Paginated Query on Primary Key``
+            testCase "Paginated Query on LSI " paginationTests.``Paginated Query on LSI``
+            testCase "Paginated Query on GSI " paginationTests.``Paginated Query on GSI``
+            testCase "Paginated Query with filter" paginationTests.``Paginated Query with filter``
+        ]
+
 [<Tests>]
 let tests =
     testList ""
@@ -211,6 +222,7 @@ let tests =
             UpdateExpressionTests
             ProjectionExpressionTests
             SparseGSITests
+            PaginationTests
         ]
 
 [<EntryPoint>]