+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+### 0.8.3-beta
+* Added `ScanPaginated*` and `QueryPaginated*` methods to `TableContext` to support paginating queries
+* Added `IndexKey` type to support additional key fields in LastEvaluatedKey for queries (ie on LSI & GSI indices)
+* **Breaking** renamed one of the method parameters from `filterExpr` to `filterCondition` for consistency
+
 ### 0.8.2-beta
 * Replace attribute name validation with something that sticks closer to the [AWS naming rules](https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/HowItWorks.NamingRulesDataTypes.html) - fixes #29
 * Update project to net50 & bumped dependencies